# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted webmail application API backed by IMAP and SMTP.",
        "title": "Go-Cubemail API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/push/subscribe": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Stores the browser PushSubscription (endpoint + keys) after the user grants notification permission.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "push"
                ],
                "summary": "Register push subscription",
                "parameters": [
                    {
                        "description": "Push subscription payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.subscribePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status subscribed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid body or missing fields",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/push/unsubscribe": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Deletes the push subscription for the given endpoint when the user revokes notification permission.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "push"
                ],
                "summary": "Remove push subscription",
                "parameters": [
                    {
                        "description": "endpoint to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status unsubscribed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/push/vapid-public-key": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns the server VAPID public key required for the browser to create a Web Push subscription.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "push"
                ],
                "summary": "Get VAPID public key",
                "responses": {
                    "200": {
                        "description": "public_key",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "push not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.subscribePayload": {
            "type": "object",
            "properties": {
                "endpoint": {
                    "type": "string"
                },
                "keys": {
                    "type": "object",
                    "properties": {
                        "auth": {
                            "type": "string"
                        },
                        "p256dh": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "The session cookie 'gorc_session' is required for authenticated endpoints.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}